zmeny

+++ hts221/hts221.c
@@ -55,7 +55,7 @@ T_out = (int16_t) ((int16_t)data[1] << 8 |	(int16_t)data[0]);	//tiez
 
 //https://www.johndcook.com/interpolatorhelp.html
 //y1 = T0_degC,y2 = T1_degC;x1 = T0Out, x2 = T1_Out, x3 = T_out
-vystupna = ((T1_out - T_out)*T0_degC + (T_out - T0_out)*T1_degC) / (T1_out -T0_out);
+vystupna = (float)((T1_out - T_out)*T0_degC + (T_out - T0_out)*T1_degC) / (T1_out -T0_out);
 if (vystupna >= 100)
 	vystupna = 99.9;
 else if (vystupna <= -100)
